change readCSV and readSQLDump to write data an output directories, fix minor bugs discovered during regression tests (= ibis1.1.3)
git-svn-id: https://codeforge.lbl.gov/anonscm/fastbit/trunk@192 9ec3b2dc-5744-0410-9993-b8723341f1db

+++ src/colValues.cpp
@@ -3029,7 +3029,7 @@ uint32_t ibis::colStrings::write(FILE* fptr) const {
     for (uint32_t i = 0; i < nelm; ++ i) {
 	int ierr = fwrite((*array)[i].c_str(), sizeof(char),
 			  (*array)[i].size()+1, fptr);
-	cnt += (int) (ierr > (*array)[i].size());
+	cnt += (int) (ierr > long((*array)[i].size()));
 	LOGGER(ierr <= 0 && ibis::gVerbose >= 0)
 	    << "Warning -- colStrings[" << col->partition()->name() << '.'
 	    << col->name() << "]::write failed to string " << (*array)[i]
